requirements.txt
Flask==3.0.3
rembg==2.0.56
Pillow==9.2.0
Flask-Cors==3.0.10